+++ fuzzer/src/cpp_interface.rs
@@ -5,7 +5,7 @@
 #[link(name = "z3")]
 extern {
   pub fn submit_task(input: *const u8, input_length: u32, fresh: bool);
-  pub fn submit_fmemcmp(data: *const u8, index: u64, size: u32, tid: u32, addr: u64);
+  pub fn submit_fmemcmp(data: *const u8, index: u32, size: u32, tid: u32, addr: u64);
   pub fn get_queue_length() -> u32;
   pub fn init_core(save_whole: bool, use_codecache: bool);
   pub fn aggregate_results();
